requirements.txt
torch
numpy
scipy
torchvision
tensorboardX
tensorflow
click
tqdm
requests
colorlog
pyyaml
